requirements.txt
matplotlib
scikit-learn==1.2.2
sklearn.ensemble
pandas
seaborn
